--- src/main/java/metro/command/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Implementation of commands used in the user interface.
- */
-@NonNullApi
-package metro.command;
-
-import org.springframework.lang.NonNullApi;
