--- src/main/java/metro/config/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Spring Configuration
- */
-@NonNullApi
-package metro.config;
-
-import org.springframework.lang.NonNullApi;
