--- src/main/java/metro/controller/package-info.java
@@ -1,7 +1,0 @@
-/**
- * This package defines a controller that processes user commands.
- */
-@NonNullApi
-package metro.controller;
-
-import org.springframework.lang.NonNullApi;
