--- src/main/java/metro/ui/UserInterface.java
@@ -1,7 +1,0 @@
-package metro.ui;
-
-public interface UserInterface {
-    String read();
-
-    void write(Object line);
-}
