--- src/main/java/metro/ui/package-info.java
@@ -1,4 +1,0 @@
-/**
- * User Interface
- */
-package metro.ui;
